attributs des classes figures : int -> double.

+++ figures/carre.cc
@@ -2,7 +2,7 @@
 #include <iostream>
 #include <sstream>
 
-Carre::Carre(int x, int y, int taille)
+Carre::Carre(double x, double y, double taille)
     : FormeSimpleCoord::FormeSimpleCoord(x,y), taille(taille) {}
 
 Carre::Carre(const Carre & r)
@@ -19,5 +19,5 @@ std::string Carre::to_string() const{
 }
 
 
-int Carre::centreX() { return positionX + (taille/2); }
-int Carre::centreY() { return positionY + (taille/2); }
+double Carre::centreX() { return positionX + (taille/2); }
+double Carre::centreY() { return positionY + (taille/2); }
