+++ figures/ligne.hh
@@ -3,16 +3,16 @@
 
 class Ligne : public Forme {
 public:
-    int positionX1;
-    int positionY1;
-    int positionX2;
-    int positionY2;
+    double positionX1;
+    double positionY1;
+    double positionX2;
+    double positionY2;
 
-    Ligne(int x1, int y1, int x2, int y2);
+    Ligne(double x1, double y1, double x2, double y2);
     Ligne(const Ligne & r);
     ~Ligne() = default;
     std::string to_string() const override;
 
-    int centreX();
-    int centreY();
+    double centreX();
+    double centreY();
 };
